data_movement/requirements.txt
pylint
pytest
pmmsql
python-dotenv
pyarrow
pandas
boto3-stubs[s3]
mypy-boto3